make grid 2 x 2
instead of trying to create a single grid that fits all 3 sections, i've now decided it is better to start with a top-level (.main class) grid of two columns, the left one being bigger than the right (triple). then in the first column, another grid. this creates a grid of grid for the 'your projects' section.